Move UuidUtils to core.
This allows it to be properly shared between ST0903 and ST1204.
Also update comments about tentative implementation - now resolved.
